Unify panic middleware to ensure panic is always logged
