+++ pkg/admin/wire.go
@@ -17,6 +17,14 @@ import (
 	"github.com/authgear/authgear-server/pkg/util/httproute"
 )
 
+func newPanicMiddleware(p *deps.RootProvider) httproute.Middleware {
+	panic(wire.Build(
+		deps.RootDependencySet,
+		middleware.DependencySet,
+		wire.Bind(new(httproute.Middleware), new(*middleware.PanicMiddleware)),
+	))
+}
+
 func newHealthzHandler(p *deps.RootProvider, w http.ResponseWriter, r *http.Request, ctx context.Context) http.Handler {
 	panic(wire.Build(
 		deps.RootDependencySet,
@@ -36,31 +44,10 @@ func newSentryMiddleware(p *deps.RootProvider) httproute.Middleware {
 	))
 }
 
-func newPanicEndMiddleware(p *deps.RootProvider) httproute.Middleware {
-	panic(wire.Build(
-		middleware.DependencySet,
-		wire.Bind(new(httproute.Middleware), new(*middleware.PanicEndMiddleware)),
-	))
-}
-
-func newPanicWriteEmptyResponseMiddleware(p *deps.RootProvider) httproute.Middleware {
-	panic(wire.Build(
-		middleware.DependencySet,
-		wire.Bind(new(httproute.Middleware), new(*middleware.PanicWriteEmptyResponseMiddleware)),
-	))
-}
-
 func newBodyLimitMiddleware(p *deps.RootProvider) httproute.Middleware {
 	panic(wire.Build(
 		middleware.DependencySet,
 		wire.Bind(new(httproute.Middleware), new(*middleware.BodyLimitMiddleware)),
-	))
-}
-
-func newPanicLogMiddleware(p *deps.RequestProvider) httproute.Middleware {
-	panic(wire.Build(
-		DependencySet,
-		wire.Bind(new(httproute.Middleware), new(*middleware.PanicLogMiddleware)),
 	))
 }
