+++ pkg/auth/wire_gen.go
@@ -25838,6 +25838,15 @@ func newWebAppWebsocketHandler(p *deps.RequestProvider) http.Handler {
 
 // Injectors from wire_middleware.go:
 
+func newPanicMiddleware(p *deps.RootProvider) httproute.Middleware {
+	factory := p.LoggerFactory
+	panicMiddlewareLogger := middleware.NewPanicMiddlewareLogger(factory)
+	panicMiddleware := &middleware.PanicMiddleware{
+		Logger: panicMiddlewareLogger,
+	}
+	return panicMiddleware
+}
+
 func newSentryMiddleware(p *deps.RootProvider) httproute.Middleware {
 	hub := p.SentryHub
 	environmentConfig := p.EnvironmentConfig
@@ -25852,31 +25861,6 @@ func newSentryMiddleware(p *deps.RootProvider) httproute.Middleware {
 func newBodyLimitMiddleware(p *deps.RootProvider) httproute.Middleware {
 	bodyLimitMiddleware := &middleware.BodyLimitMiddleware{}
 	return bodyLimitMiddleware
-}
-
-func newPanicEndMiddleware(p *deps.RootProvider) httproute.Middleware {
-	panicEndMiddleware := &middleware.PanicEndMiddleware{}
-	return panicEndMiddleware
-}
-
-func newPanicWriteEmptyResponseMiddleware(p *deps.RequestProvider) httproute.Middleware {
-	panicWriteEmptyResponseMiddleware := &middleware.PanicWriteEmptyResponseMiddleware{}
-	return panicWriteEmptyResponseMiddleware
-}
-
-func newPanicLogMiddleware(p *deps.RequestProvider) httproute.Middleware {
-	appProvider := p.AppProvider
-	factory := appProvider.LoggerFactory
-	panicLogMiddlewareLogger := middleware.NewPanicLogMiddlewareLogger(factory)
-	panicLogMiddleware := &middleware.PanicLogMiddleware{
-		Logger: panicLogMiddlewareLogger,
-	}
-	return panicLogMiddleware
-}
-
-func newPanicAPIMiddleware(p *deps.RequestProvider) httproute.Middleware {
-	panicWriteAPIResponseMiddleware := &middleware.PanicWriteAPIResponseMiddleware{}
-	return panicWriteAPIResponseMiddleware
 }
 
 func newPanicWebAppMiddleware(p *deps.RequestProvider) httproute.Middleware {
