+++ pkg/lib/infra/middleware/deps.go
@@ -8,13 +8,8 @@ var DependencySet = wire.NewSet(
 	NewCORSMiddlewareLogger,
 	wire.Struct(new(CORSMiddleware), "*"),
 
-	NewPanicLogMiddlewareLogger,
-	wire.Struct(new(PanicLogMiddleware), "*"),
-
-	wire.Struct(new(PanicWriteAPIResponseMiddleware), "*"),
-	wire.Struct(new(PanicWriteEmptyResponseMiddleware), "*"),
-
-	wire.Struct(new(PanicEndMiddleware), "*"),
+	NewPanicMiddlewareLogger,
+	wire.Struct(new(PanicMiddleware), "*"),
 
 	wire.Struct(new(SentryMiddleware), "*"),
