+++ pkg/portal/wire.go
@@ -14,25 +14,11 @@ import (
 	"github.com/authgear/authgear-server/pkg/util/httproute"
 )
 
-func newPanicEndMiddleware(p *deps.RequestProvider) httproute.Middleware {
-	panic(wire.Build(
-		middleware.DependencySet,
-		wire.Bind(new(httproute.Middleware), new(*middleware.PanicEndMiddleware)),
-	))
-}
-
-func newPanicLogMiddleware(p *deps.RequestProvider) httproute.Middleware {
+func newPanicMiddleware(p *deps.RequestProvider) httproute.Middleware {
 	panic(wire.Build(
 		deps.DependencySet,
 		middleware.DependencySet,
-		wire.Bind(new(httproute.Middleware), new(*middleware.PanicLogMiddleware)),
-	))
-}
-
-func newPanicWriteEmptyResponseMiddleware(p *deps.RequestProvider) httproute.Middleware {
-	panic(wire.Build(
-		middleware.DependencySet,
-		wire.Bind(new(httproute.Middleware), new(*middleware.PanicWriteEmptyResponseMiddleware)),
+		wire.Bind(new(httproute.Middleware), new(*middleware.PanicMiddleware)),
 	))
 }
